@@ -15,7 +15,7 @@ class ApiKeyUser implements UserInterface
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column(type: 'integer')]
-    private $id;
+    private ?int $id;
 
     #[ORM\Column(type: 'string', length: 255, unique: true)]
     #[Assert\Length(
@@ -42,6 +42,9 @@ class ApiKeyUser implements UserInterface
         return self::ROLES;
     }
 
+    /**
+     * @return void
+     */
     public function eraseCredentials()
     {
     }
