@@ -3,6 +3,7 @@
 namespace App\Command;
 
 use App\Service\ApiKeyUserService;
+use Exception;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,28 +16,23 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 )]
 class AuthCreateApikeyCommand extends Command
 {
-    public function __construct(private ApiKeyUserService $apiKeyUserService)
-    {
+    public function __construct(
+        private readonly ApiKeyUserService $apiKeyUserService
+    ) {
         parent::__construct();
     }
 
-    protected function configure(): void
-    {
-    }
-
     /**
-     * @throws \Exception
+     * @throws Exception
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
 
         $name = $io->ask('Enter name for ApiKeyUser');
-
         $webformApiKey = $io->ask('Enter webform api-key');
 
         $apiKeyUser = $this->apiKeyUserService->createApiKey($name, $webformApiKey);
-
         $apiKey = $apiKeyUser->getApiKey();
 
         $io->writeln('------------');
