@@ -16,29 +16,25 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 )]
 class GraphUserBookingsCommand extends Command
 {
-    public function __construct(private MicrosoftGraphServiceInterface $microsoftGraphService)
-    {
+    public function __construct(
+        private readonly MicrosoftGraphServiceInterface $microsoftGraphService
+    ) {
         parent::__construct();
     }
 
     protected function configure(): void
     {
-        $this->addArgument(
-            'userId',
-            InputArgument::REQUIRED,
-            'User id to get bookings for.'
-        );
+        $this->addArgument('userId', InputArgument::REQUIRED, 'User id to get bookings for.');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
-        $userId = $input->getArgument('userId');
 
+        $userId = $input->getArgument('userId');
         $io->note(sprintf('You request bookings for userId: %s', $userId));
 
         $data = $this->microsoftGraphService->getUserBookings($userId);
-
         $io->info(json_encode($data));
 
         return Command::SUCCESS;
