@@ -12,8 +12,9 @@ use Symfony\Component\Uid\Ulid;
 
 final class BusyIntervalCollectionDataProvider implements ContextAwareCollectionDataProviderInterface, RestrictedDataProviderInterface
 {
-    public function __construct(private MicrosoftGraphServiceInterface $microsoftGraphService)
-    {
+    public function __construct(
+        private readonly MicrosoftGraphServiceInterface $microsoftGraphService
+    ) {
     }
 
     public function supports(string $resourceClass, string $operationName = null, array $context = []): bool
