@@ -7,17 +7,21 @@ use ApiPlatform\Core\DataProvider\RestrictedDataProviderInterface;
 use App\Entity\Main\UserBooking;
 use App\Security\Voter\UserBookingVoter;
 use App\Service\MicrosoftGraphServiceInterface;
-use Exception;
 use GuzzleHttp\Exception\GuzzleException;
 use Microsoft\Graph\Exception\GraphException;
+use Psr\Cache\InvalidArgumentException;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\Security\Core\Security;
 use Symfony\Component\Uid\Ulid;
 
 final class UserBookingCollectionDataProvider implements ContextAwareCollectionDataProviderInterface, RestrictedDataProviderInterface
 {
-    public function __construct(private MicrosoftGraphServiceInterface $microsoftGraphService, private Security $security)
-    {
+    public function __construct(
+        private readonly MicrosoftGraphServiceInterface $microsoftGraphService,
+        private readonly Security $security,
+        private readonly RequestStack $requestStack
+    ) {
     }
 
     public function supports(string $resourceClass, string $operationName = null, array $context = []): bool
@@ -26,29 +30,30 @@ final class UserBookingCollectionDataProvider implements ContextAwareCollectionD
     }
 
     /**
-     * @throws GuzzleException
      * @throws GraphException
-     * @throws Exception
+     * @throws InvalidArgumentException
+     * @throws GuzzleException
      */
     public function getCollection(string $resourceClass, string $operationName = null, array $context = []): iterable
     {
-        if (!isset($context['filters'])) {
-            throw new BadRequestHttpException('Required filters not set.');
+        $request = $this->requestStack->getCurrentRequest();
+
+        if (is_null($request)) {
+            throw new BadRequestHttpException('Request not set.');
         }
 
-        $filters = $context['filters'];
+        $userId = $request->headers->get('Authorization-UserId') ?? null;
 
-        if (!isset($filters['userId'])) {
-            throw new BadRequestHttpException('Required userId filter not set.');
+        if (is_null($userId)) {
+            throw new BadRequestHttpException('Required Authorization-UserId header is not set.');
         }
-
-        $userId = $filters['userId'];
 
         $userBookings = $this->microsoftGraphService->getUserBookings($userId);
+
         $userBookingsHits = $userBookings['value'][0]['hitsContainers'][0]['hits'] ?? null;
 
         if (null === $userBookingsHits) {
-            return 'no results';
+            return;
         }
 
         foreach ($userBookingsHits as $hit) {
@@ -58,8 +63,7 @@ final class UserBookingCollectionDataProvider implements ContextAwareCollectionD
             $userBooking->subject = $hit['resource']['subject'] ?? '';
             $userBooking->start = new \DateTime($hit['resource']['start']['dateTime'], new \DateTimeZone($hit['resource']['start']['timeZone'])) ?? null;
             $userBooking->end = new \DateTime($hit['resource']['end']['dateTime'], new \DateTimeZone($hit['resource']['end']['timeZone'])) ?? null;
-            // TODO: Is the summary needed? Atm. it is hidden.
-            // $userBooking->summary = $hit['summary'] ?? '';
+            $userBooking->summary = $hit['summary'] ?? '';
 
             $bookingDetailsData = $this->microsoftGraphService->getBookingDetails($hit['hitId']);
 
@@ -67,10 +71,6 @@ final class UserBookingCollectionDataProvider implements ContextAwareCollectionD
             $userBooking->body = $bookingDetailsData['body']['content'];
 
             if ($this->security->isGranted(UserBookingVoter::VIEW, $userBooking)) {
-                // TODO: Is the body needed? Atm. it is hidden.
-                // The body is needed for the security voter, but should not be delivered to the user.
-                // TODO: Handle this with a DTO and data transformer.
-                unset($userBooking->body);
                 yield $userBooking;
             }
         }
