@@ -13,17 +13,20 @@ use Symfony\Component\Security\Core\Security;
 
 final class AAKResourceExtension implements QueryCollectionExtensionInterface, QueryItemExtensionInterface
 {
-    public function __construct(private Security $security, private RequestStack $requestStack, private CvrWhitelistRepository $cvrWhitelistRepository)
-    {
+    public function __construct(
+        private readonly Security $security,
+        private readonly RequestStack $requestStack,
+        private readonly CvrWhitelistRepository $cvrWhitelistRepository
+    ) {
     }
 
-    public function applyToCollection(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null)
+    public function applyToCollection(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null): void
     {
         $this->applyResourceRequireLocation($queryBuilder, $resourceClass);
         $this->applyWhitelistPermission($queryBuilder, $resourceClass);
     }
 
-    public function applyToItem(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, array $identifiers, string $operationName = null, array $context = [])
+    public function applyToItem(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, array $identifiers, string $operationName = null, array $context = []): void
     {
         $this->applyResourceRequireLocation($queryBuilder, $resourceClass);
         $this->applyWhitelistPermission($queryBuilder, $resourceClass);
@@ -52,7 +55,7 @@ final class AAKResourceExtension implements QueryCollectionExtensionInterface, Q
 
         // Extract whitelistKey from request.
         $currentRequest = $this->requestStack->getCurrentRequest();
-        $whitelistKey = $currentRequest->query->get('whitelistKey');
+        $whitelistKey = $currentRequest?->query->get('whitelistKey');
 
         // If whitelistKey is set, check if the whitelistKey exists in cvrWhitelist for the given resource.
         if (null !== $whitelistKey) {
