@@ -12,7 +12,6 @@ use App\Service\WebformServiceInterface;
 use App\Utils\ValidationUtilsInterface;
 use DateTime;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
 use Symfony\Component\Messenger\MessageBusInterface;
@@ -28,12 +27,12 @@ use Twig\Error\SyntaxError;
 class WebformSubmitHandler
 {
     public function __construct(
-        private WebformServiceInterface $webformService,
-        private MessageBusInterface $bus,
-        private ValidationUtilsInterface $validationUtils,
-        private LoggerInterface $logger,
-        private AAKResourceRepository $aakResourceRepository,
-        private Environment $twig,
+        private readonly WebformServiceInterface $webformService,
+        private readonly MessageBusInterface $bus,
+        private readonly ValidationUtilsInterface $validationUtils,
+        private readonly LoggerInterface $logger,
+        private readonly AAKResourceRepository $aakResourceRepository,
+        private readonly Environment $twig,
     ) {
     }
 
@@ -83,16 +82,9 @@ class WebformSubmitHandler
     }
 
     /**
-     * @param $data
-     * @param $email
-     * @param $resource
-     * @param $metaData
-     *
-     * @return array
-     *
-     * @throws Exception|\Exception
+     * @throws \Exception
      */
-    private function composeBookingContents($data, $email, $resource, $metaData): array
+    private function composeBookingContents($data, string $email, $resource, $metaData): array
     {
         $body = [];
 
@@ -110,15 +102,11 @@ class WebformSubmitHandler
     }
 
     /**
-     * @param $body
-     *
-     * @return string
-     *
-     * @throws LoaderError
      * @throws RuntimeError
      * @throws SyntaxError
+     * @throws LoaderError
      */
-    private function renderContentsAsHtml($body): string
+    private function renderContentsAsHtml(array $body): string
     {
         return $this->twig->render('booking.html.twig', $body);
     }
