@@ -7,18 +7,12 @@ use App\Repository\Resources\CvrWhitelistRepository;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 
-/**
- * @extends ServiceEntityRepository<AakResource>
- *
- * @method resource|null find($id, $lockMode = null, $lockVersion = null)
- * @method resource|null findOneBy(array $criteria, array $orderBy = null)
- * @method resource[]    findAll()
- * @method resource[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
 class AAKResourceRepository extends ServiceEntityRepository
 {
-    public function __construct(ManagerRegistry $registry, private CvrWhitelistRepository $cvrWhitelistRepository)
-    {
+    public function __construct(
+        ManagerRegistry $registry,
+        private readonly CvrWhitelistRepository $cvrWhitelistRepository
+    ) {
         parent::__construct($registry, AAKResource::class);
     }
 
@@ -31,7 +25,7 @@ class AAKResourceRepository extends ServiceEntityRepository
         }
     }
 
-    public function findOneByEmail(string $email): ?AAKResource
+    public function findOneByEmail(string $email)
     {
         return $this->findOneBy(['resourceMail' => $email]);
     }
