@@ -28,8 +28,13 @@ class ApiKeyAuthenticator extends AbstractAuthenticator
      */
     public function supports(Request $request): ?bool
     {
-        return $request->headers->has(self::AUTH_HEADER) &&
-            str_starts_with($request->headers->get(self::AUTH_HEADER), self::AUTH_HEADER_PREFIX);
+        if ($request->headers->has(self::AUTH_HEADER)) {
+            $authHeader = $request->headers->get(self::AUTH_HEADER);
+
+            return !is_null($authHeader) && str_starts_with($authHeader, self::AUTH_HEADER_PREFIX);
+        }
+
+        return false;
     }
 
     /**
